// frontend/src/generated/Effects.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChorusEffect } from "./ChorusEffect";
import type { DelayEffect } from "./DelayEffect";
import type { DistortionEffect } from "./DistortionEffect";
import type { EqEffect } from "./EqEffect";
import type { ReverbEffect } from "./ReverbEffect";

export type Effects = { eq?: EqEffect, distortion?: DistortionEffect, chorus?: ChorusEffect, delay?: DelayEffect, reverb?: ReverbEffect, };